feat(commandsModule): Add setBrushSize and setThresholdRange actions for segmentation tools
- Implemented setBrushSize action to adjust brush size for specified tool groups.
- Added setThresholdRange action to configure threshold ranges for various segmentation tools.
- Removed obsolete setBrushSize and setThresholdRange actions from cornerstone-dicom-seg commands module.
